creditagent/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
streamlit==1.32.0
xgboost==2.0.3
shap==0.44.1
anthropic==0.21.0
google-genai>=1.0.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
python-dotenv==1.0.0
httpx==0.27.0
plotly==5.20.0
